requirements.txt
uvicorn
fastapi
boto3
python-dotenv
pydantic
requests
google-genai
pytrends
